refactor: fix test
